+++ src/values/color/cmyk_color.nim
@@ -1,16 +1,16 @@
 type CmykColor* = object
-  c*: uint
-  m*: uint
-  y*: uint
-  k*: uint
+  c*: float
+  m*: float
+  y*: float
+  k*: float
 
-proc getCyanColor*(cmyk: CmykColor): uint = cmyk.c
+proc getCyanColor*(cmyk: CmykColor): float = cmyk.c
 
-proc getMagentaColor*(cmyk: CmykColor): uint = cmyk.m
+proc getMagentaColor*(cmyk: CmykColor): float = cmyk.m
 
-proc getYellowColor*(cmyk: CmykColor): uint =  cmyk.y
+proc getYellowColor*(cmyk: CmykColor): float =  cmyk.y
 
-proc getBlackColor*(cmyk: CmykColor): uint =  cmyk.k
+proc getBlackColor*(cmyk: CmykColor): float =  cmyk.k
 
-proc createCmykColor*(c: uint, m: uint, y: uint, k: uint): CmykColor = 
+proc createCmykColor*(c: float, m: float, y: float, k: float): CmykColor = 
   return CmykColor(c: c, m: m, y: y, k: k)
